+++ tomcat/src/test/java/org/apache/coyote/http11/request/http/ProtocolTest.java
@@ -1,4 +1,4 @@
-package nextstep.jwp.protocol.http;
+package org.apache.coyote.http11.request.http;
 
 import org.apache.coyote.http11.request.line.Protocol;
 import org.junit.jupiter.api.Assertions;
